# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Berlin MarketHUB API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: Get reviews (for a listing or seller)
      parameters:
        - name: listingId
          in: query
          schema:
            type: string
        - name: sellerId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /upload/image:
    post:
      operationId: processImageUpload
      tags: [upload]
      summary: Process and compress an uploaded image URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUploadInput"
      responses:
        "200":
          description: Processed image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
...
components:
  schemas:
...
    Review:
      type: object
      properties:
        id:
          type: string
        listingId:
          type: ["string", "null"]
        sellerId:
          type: string
        reviewerId:
          type: string
        reviewerName:
          type: string
        reviewerAvatar:
          type: ["string", "null"]
        rating:
          type: integer
        comment:
          type: string
        isVerifiedPurchase:
          type: boolean
        helpfulCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - sellerId
        - reviewerId
        - reviewerName
        - rating
        - comment
        - isVerifiedPurchase
        - helpfulCount
        - createdAt
    ReviewInput:
      type: object
      properties:
        listingId:
          type: ["string", "null"]
        sellerId:
          type: string
        rating:
          type: integer
        comment:
          type: string
      required:
        - sellerId
        - rating
        - comment
...
    ImageUploadInput:
      type: object
      properties:
        dataUrl:
          type: string
        bucket:
          type: string
        filename:
          type: string
      required:
        - dataUrl
    UploadResult:
      type: object
      properties:
        url:
          type: string
        path:
          type: string
        size:
          type: integer
      required:
        - url
        - path
        - size
...